+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix sliderWidth when there was less items than the number of itens to show per slide.
+
 ## [1.11.0] - 2019-04-02
 
 ### Added